# movies/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  '/api/movies/{id}':
    get:
      summary: ''
      operationId: getApiMoviesId
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
      security: []
    delete:
      summary: ''
      operationId: deleteApiMoviesId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    patch:
      summary: ''
      operationId: patchApiMoviesId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: vuxcqufoymyyfkuuj
                genre:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: unbzdbxnveipejlvar
                description:
                  type: string
                  description: ''
                  example: 'Pariatur repellat animi sapiente.'
                director:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: dfpzdulwmxadkkgsfrdhyb
                year:
                  type: integer
                  description: ''
                  example: 14
      security: []
    put:
      summary: ''
      operationId: putApiMoviesId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: rcooqxtmkeycbgefuv
                genre:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: nghonih
                description:
                  type: string
                  description: ''
                  example: 'Odit quod reprehenderit ex et.'
                director:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: mhnvvxasxyfrjfgzsitw
                year:
                  type: integer
                  description: ''
                  example: 13
              required:
                - title
                - genre
                - description
                - director
                - year
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the movie.'
        example: sequi
        required: true
        schema:
          type: string
...
  /api/movies:
    post:
      summary: ''
      operationId: postApiMovies
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: ogfxftlpuodqrpvh
                genre:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: hptmsggtos
                description:
                  type: string
                  description: ''
                  example: 'Voluptate provident placeat et id eos rerum.'
                director:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: dog
                year:
                  type: integer
                  description: ''
                  example: 4
              required:
                - title
                - genre
                - description
                - director
                - year
      security: []
...